fix: error is shown for valid emails
In case user enters incorrect email, we show an error. After that, if the user fixes the error (for example has a space in the beginning), the default email validator does not receive the new value.
To fix this use a custom validation with the same regex used inside the EmailValidator. Also, trim the email before the validation.